+++ tests/test_header_parameter.py
@@ -177,5 +177,5 @@ def test_get_header_advanced_configuration(
     assert headers["X-Pxl-Envvar"] == os.environ["USERNAME"]
     assert headers["X-Pxl-Request"]
     assert headers["User-Agent"] == f"PyxelRest v{__version__}"
-    assert headers["X-Pxl-Cell"] == ""
+    assert headers["X-Pxl-Caller"] == ""
     assert headers["X-Pxl-Session"] == "2018-10-11T15:05:05.663979"